refactor: separate service.go into multiple files based on the domain
