+++ pkg/service/secret.go
@@ -1,0 +1,98 @@
+package service
+
+import (
+	"context"
+
+	"go.einride.tech/aip/filtering"
+
+	"github.com/instill-ai/pipeline-backend/internal/resource"
+
+	pb "github.com/instill-ai/protogen-go/vdp/pipeline/v1beta"
+)
+
+func (s *service) CreateNamespaceSecret(ctx context.Context, ns resource.Namespace, pbSecret *pb.Secret) (*pb.Secret, error) {
+
+	if err := s.checkNamespacePermission(ctx, ns); err != nil {
+		return nil, err
+	}
+
+	dbSecret, err := s.PBToDBSecret(ctx, ns, pbSecret)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := s.repository.CreateNamespaceSecret(ctx, ns.Permalink(), dbSecret); err != nil {
+		return nil, err
+	}
+
+	dbCreatedSecret, err := s.repository.GetNamespaceSecretByID(ctx, ns.Permalink(), dbSecret.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	return s.DBToPBSecret(ctx, dbCreatedSecret)
+
+}
+
+func (s *service) ListNamespaceSecrets(ctx context.Context, ns resource.Namespace, pageSize int32, pageToken string, filter filtering.Filter) ([]*pb.Secret, int32, string, error) {
+
+	if err := s.checkNamespacePermission(ctx, ns); err != nil {
+		return nil, 0, "", err
+	}
+
+	dbSecrets, ps, pt, err := s.repository.ListNamespaceSecrets(ctx, ns.Permalink(), int64(pageSize), pageToken, filter)
+	if err != nil {
+		return nil, 0, "", err
+	}
+
+	pbSecrets, err := s.DBToPBSecrets(ctx, dbSecrets)
+	return pbSecrets, int32(ps), pt, err
+}
+
+func (s *service) GetNamespaceSecretByID(ctx context.Context, ns resource.Namespace, id string) (*pb.Secret, error) {
+
+	if err := s.checkNamespacePermission(ctx, ns); err != nil {
+		return nil, err
+	}
+
+	ownerPermalink := ns.Permalink()
+
+	dbSecret, err := s.repository.GetNamespaceSecretByID(ctx, ownerPermalink, id)
+	if err != nil {
+		return nil, ErrNotFound
+	}
+	return s.DBToPBSecret(ctx, dbSecret)
+}
+
+func (s *service) UpdateNamespaceSecretByID(ctx context.Context, ns resource.Namespace, id string, updatedSecret *pb.Secret) (*pb.Secret, error) {
+
+	if err := s.checkNamespacePermission(ctx, ns); err != nil {
+		return nil, err
+	}
+
+	ownerPermalink := ns.Permalink()
+
+	dbSecret, err := s.PBToDBSecret(ctx, ns, updatedSecret)
+	if err != nil {
+		return nil, ErrNotFound
+	}
+
+	if _, err = s.repository.GetNamespaceSecretByID(ctx, ownerPermalink, id); err != nil {
+		return nil, err
+	}
+
+	if err := s.repository.UpdateNamespaceSecretByID(ctx, ns.Permalink(), id, dbSecret); err != nil {
+		return nil, err
+	}
+
+	return s.GetNamespaceSecretByID(ctx, ns, id)
+}
+
+func (s *service) DeleteNamespaceSecretByID(ctx context.Context, ns resource.Namespace, id string) error {
+	if err := s.checkNamespacePermission(ctx, ns); err != nil {
+		return err
+	}
+	ownerPermalink := ns.Permalink()
+
+	return s.repository.DeleteNamespaceSecretByID(ctx, ownerPermalink, id)
+}
